[cms-sanity] Move server-only utilities into separate file
