+++ examples/cms-sanity/lib/sanity.js
@@ -1,47 +1,14 @@
 import {
-  createClient,
   createImageUrlBuilder,
   createPreviewSubscriptionHook,
 } from 'next-sanity'
+import { sanityConfig } from './config'
 
-const options = {
-  // Find your project ID and dataset in `sanity.json` in your studio project
-  dataset: process.env.NEXT_PUBLIC_SANITY_DATASET || 'production',
-  projectId: process.env.NEXT_PUBLIC_SANITY_PROJECT_ID,
-  useCdn: process.env.NODE_ENV === 'production',
-  // useCdn == true gives fast, cheap responses using a globally distributed cache.
-  // Set this to false if your application require the freshest possible
-  // data always (potentially slightly slower and a bit more expensive).
-}
-
-export const sanityClient = createClient(options)
-
-export const previewClient = createClient({
-  ...options,
-  useCdn: false,
-  token: process.env.SANITY_API_TOKEN,
-})
-
-export const getClient = (preview) => (preview ? previewClient : sanityClient)
-
-export const imageBuilder = createImageUrlBuilder(options)
+export const imageBuilder = createImageUrlBuilder(sanityConfig)
 
 export const urlForImage = (source) =>
   imageBuilder.image(source).auto('format').fit('max')
 
-export const usePreviewSubscription = createPreviewSubscriptionHook(options)
-
-export function overlayDrafts(docs) {
-  const documents = docs || []
-  const overlayed = documents.reduce((map, doc) => {
-    if (!doc._id) {
-      throw new Error('Ensure that `_id` is included in query projection')
-    }
-
-    const isDraft = doc._id.startsWith('drafts.')
-    const id = isDraft ? doc._id.slice(7) : doc._id
-    return isDraft || !map.has(id) ? map.set(id, doc) : map
-  }, new Map())
-
-  return Array.from(overlayed.values())
-}
+export const usePreviewSubscription = createPreviewSubscriptionHook(
+  sanityConfig
+)
